Move create_proof_by_ids into NonFungibleVault
