+++ radix-engine/src/blueprints/resource/resource_manager_util.rs
@@ -187,7 +187,7 @@ fn build_access_rules(
         DenyAll,
     );
     vault_access_rules.set_method_access_rule_and_mutability(
-        MethodKey::new(NodeModuleId::SELF, VAULT_CREATE_PROOF_BY_IDS_IDENT),
+        MethodKey::new(NodeModuleId::SELF, NON_FUNGIBLE_VAULT_CREATE_PROOF_BY_IDS_IDENT),
         AllowAll,
         DenyAll,
     );
